move Swift to own dir
