+++ packages/quicktype-core/src/language/Swift/index.ts
@@ -1,0 +1,2 @@
+export { SwiftTargetLanguage, swiftOptions } from "./language";
+export { SwiftRenderer } from "./SwiftRenderer";
